# api.yaml
openapi: 3.0.3
info:
  title: Viking Demo API
  version: 1.0.0
  description: Демонстрационное REST API для получения списка созданных викингов

servers:
  - url: http://localhost:8081
    description: Локальный сервер разработки

paths:
  /api/vikings:
    get:
      tags:
        - Vikings
      summary: Получить список всех созданных викингов
      description: Возвращает список всех викингов, созданных в приложении на текущий момент
      operationId: getAllVikings
      responses:
        '200':
          description: Список викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viking'
    post:
      tags:
        - Vikings
      summary: Создать конкретного викинга
      description: Создать нового викинга с заданными параметрами
      operationId: addViking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viking'
      responses:
        '200':
          description: Викинг успешно создан
    put:
      tags:
        - Vikings
      summary: Обновить сведения викинга
      description: Обновить данные существующего викинга
      operationId: updateViking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Viking'
      responses:
        '200':
          description: Сведения о викинге обновлены
    delete:
      tags:
        - Vikings
      summary: Удалить викинга из таблицы
      description: Удалить викинга по его имени
      operationId: deleteViking
      parameters:
        - name: vikingName
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Викинг удален

  /api/vikings/test:
    get:
      tags:
        - Vikings
      summary: Получить список тестовых викингов
      description: Получить список тестовых викингов
      operationId: getTest
      responses:
        '200':
          description: Список викингов успешно получен
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: string

  /api/vikings/post:
    post:
      tags:
        - Vikings
      summary: Создать викинга со случайными параметрами
      description: Создать викинга со случайными параметрами
      operationId: post
      responses:
        '200':
          description: Викинг успешно создан

  /api/vikings/stats/count-by-age:
    get:
      tags:
        - Count
      summary: Подсчитать викингов по возрасту
      description: Подсчитать количество викингов, соответствующих заданным критериям возраста
      operationId: getCountByAge
      parameters:
        - name: lessThan
          in: query
          schema:
            type: integer
        - name: greaterThan
          in: query
          schema:
            type: integer
        - name: inRangeStart
          in: query
          schema:
            type: integer
        - name: inRangeEnd
          in: query
          schema:
            type: integer
        - name: outOfRangeStart
          in: query
          schema:
            type: integer
        - name: outOfRangeEnd
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Количество успешно получено
          content:
            application/json:
              schema:
                type: integer
                format: int64

  /api/vikings/stats/count-by-appearance:
    get:
      tags:
        - Count
      summary: Подсчитать викингов по внешности
      description: Подсчитать количество викингов с определенным стилем бороды или цветом волос
      operationId: getCountByAppearance
      parameters:
        - name: beardStyle
          in: query
          schema:
            $ref: '#/components/schemas/BeardStyle'
        - name: hairColor
          in: query
          schema:
            $ref: '#/components/schemas/HairColor'
      responses:
        '200':
          description: Количество успешно получено
          content:
            application/json:
              schema:
                type: integer
                format: int64

  /api/vikings/stats/count-by-axes:
    get:
      tags:
        - Count
      summary: Подсчитать викингов с 1 или 2 топорами
      description: Подсчитать количество викингов, имеющих в снаряжении один или два топора
      operationId: getCountByAxes
      responses:
        '200':
          description: Количество успешно получено
          content:
            application/json:
              schema:
                type: integer
                format: int64

  /api/vikings/ids/max:
    get:
      tags:
        - IDs
      summary: Найти последнюю запись (max ID)
      description: Найти и вернуть викинга с максимальным идентификатором
      operationId: getMaxIdViking
      responses:
        '200':
          description: Запись успешно найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Viking'

  /api/vikings/ids/even:
    get:
      tags:
        - IDs
      summary: Получить все записи с четными ID
      description: Получить список всех викингов, имеющих четные идентификаторы
      operationId: getEvenIdVikings
      responses:
        '200':
          description: Список успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Viking'

components:
  schemas:
    Viking:
      type: object
      required:
        - name
        - age
        - heightCm
        - hairColor
        - beardStyle
        - equipment
      properties:
        name:
          type: string
          example: Ragnar
        age:
          type: integer
          format: int32
          example: 34
        heightCm:
          type: integer
          format: int32
          example: 182
        hairColor:
          $ref: '#/components/schemas/HairColor'
        beardStyle:
          $ref: '#/components/schemas/BeardStyle'
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentItem'

    EquipmentItem:
      type: object
      required:
        - name
        - quality
      properties:
        name:
          type: string
          example: Axe
        quality:
          $ref: '#/components/schemas/EquipmentQuality'

    HairColor:
      type: string
      enum:
        - Blond
        - Red
        - Brown
        - Black
        - Gray
      example: Blond

    BeardStyle:
      type: string
      enum:
        - Short
        - Long
        - Braided
        - Forked
        - None
      example: Braided

    EquipmentQuality:
      type: string
      enum:
        - Common
        - Uncommon
        - Rare
        - Legendary
      example: Rare